# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "To Do List API documentation",
        "description": "To Do List OpenAPI/Swagger spec",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "example@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Returns a List of Tasks",
                "description": "Returns a List of Task Resources",
                "operationId": "getTasks",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Create a Task",
                "description": "Create a single task",
                "operationId": "addTask",
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/tasks/{taskId}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Find Task by ID",
                "description": "Return a single task",
                "operationId": "getTaskById",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Task id to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "tasks"
                ],
                "summary": "Deletes a task",
                "description": "Remove the specified resource from storage.",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Task id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful delete task"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "tasks"
                ],
                "summary": "Update the specified resource in storage.",
                "operationId": "updateTask",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "tasks",
            "description": "Task Resource"
        }
    ],
    "security": [
        []
    ]
}